Refactor exploit functionality into dedicated package
