+++ client/cmd/exploit.go
@@ -2,17 +2,15 @@
 package cmd
 
 import (
-	"errors"
 	"fmt"
 	"os"
-	"os/exec"
 	"os/signal"
 	"path/filepath"
 	"syscall"
 
-	"github.com/ByteTheCookies/cookieclient/internal/api"
 	"github.com/ByteTheCookies/cookieclient/internal/config"
 	"github.com/ByteTheCookies/cookieclient/internal/executor"
+	"github.com/ByteTheCookies/cookieclient/internal/exploit"
 	"github.com/ByteTheCookies/cookieclient/internal/filesystem"
 	"github.com/ByteTheCookies/cookieclient/internal/logger"
 	"github.com/ByteTheCookies/cookieclient/internal/submitter"
@@ -50,7 +48,7 @@ var listCmd = &cobra.Command{
 	Use:   "list",
 	Short: "List all current running exploits",
 	Long:  "This command lists all currently running exploits that are registered in the local configuration.",
-	Run:   ListExploit,
+	Run:   List,
 }
 
 // removeCmd represents the remove exploit command
@@ -66,7 +64,7 @@ var stopCmd = &cobra.Command{
 	Use:   "stop",
 	Short: "Stop the exploit execution",
 	Long:  `This command stops the execution of the currently running exploit. It gracefully terminates the exploit process and cleans up any resources used during execution.`,
-	Run:   stopExploit,
+	Run:   Stop,
 }
 
 // Common variables
@@ -112,77 +110,6 @@ func init() {
 
 // ===== RUN EXPLOIT FUNCTIONS =====
 
-// SetupClient handles the full initialization process for the attack command
-func setupClient() error {
-	var err error
-
-	err = config.LoadLocalConfig()
-	if err != nil {
-		logger.Log.Error().Err(err).Msg("Error loading local configuration, try running `cookieclient config reset`")
-		return err
-	}
-
-	if config.ArgsAttackInstance.Detach {
-		fmt.Println(logger.Blue + "[INFO]" + logger.Reset + " | Detaching from terminal")
-		Detach()
-	}
-
-	config.ArgsAttackInstance.ExploitPath, err = filesystem.NormalizeNamePathExploit(config.ArgsAttackInstance.ExploitPath)
-	if err != nil {
-		logger.Log.Error().Err(err).Msg("Error normalizing exploit name")
-		return err
-	}
-
-	if !filesystem.IsPath(config.ArgsAttackInstance.ExploitPath) {
-		exploitsDir := filepath.Join(config.DefaultConfigPath, "exploits")
-		config.ArgsAttackInstance.ExploitPath = filepath.Join(exploitsDir, config.ArgsAttackInstance.ExploitPath)
-	}
-
-	logger.Log.Debug().Str("Exploit path", config.ArgsAttackInstance.ExploitPath).Msg("Using default exploit path")
-
-	err = ValidateArgs(config.ArgsAttackInstance)
-	if err != nil {
-		return fmt.Errorf("invalid arguments: %w", err)
-	}
-
-	logger.Log.Debug().
-		Int("ThreadCount", config.ArgsAttackInstance.ThreadCount).
-		Int("Tick time", config.ArgsAttackInstance.TickTime).
-		Str("ExploitPath", config.ArgsAttackInstance.ExploitPath).
-		Msg("Arguments validated")
-
-	config.Token, err = config.GetSession()
-	if err != nil {
-		return fmt.Errorf("failed to get session token: %w, try to run `cookieclient config login -P <password>`", err)
-	}
-
-	config.Current, err = api.GetConfig()
-	if err != nil {
-		return fmt.Errorf("failed to get config: %w", err)
-	}
-
-	logger.Log.Debug().Msgf("Current configuration: %+v", config.Current)
-
-	if !config.Current.Configured {
-		logger.Log.Fatal().Msg("Server not configured. Please run the configurator before using the client.")
-	}
-
-	return nil
-}
-
-// cleanupExploit removes an exploit from the configuration
-func cleanupExploit(pid int) {
-	for i, exploit := range config.ArgsConfigInstance.Exploits {
-		if exploit.PID == pid {
-			config.ArgsConfigInstance.Exploits = append(config.ArgsConfigInstance.Exploits[:i], config.ArgsConfigInstance.Exploits[i+1:]...)
-			break
-		}
-	}
-	if err := config.WriteConfig(); err != nil {
-		logger.Log.Error().Err(err).Msg("Failed to write configuration after exploit cleanup")
-	}
-}
-
 // attack is the main execution flow for the attack command
 func attack(cmd *cobra.Command, args []string) {
 	c := make(chan os.Signal, 1)
@@ -191,12 +118,12 @@ func attack(cmd *cobra.Command, args []string) {
 		<-c
 		logger.Log.Info().Msg("Shutting down gracefully...")
 		if pid != 0 {
-			cleanupExploit(pid)
+			exploit.Cleanup(pid)
 		}
 		os.Exit(0)
 	}()
 
-	if err := setupClient(); err != nil {
+	if err := exploit.Setup(); err != nil {
 		if logger.LogLevel != zerolog.Disabled {
 			logger.Log.Fatal().Err(err).Msg("Initialization error")
 			logger.Close()
@@ -214,11 +141,11 @@ func attack(cmd *cobra.Command, args []string) {
 		logger.Log.Fatal().Err(err).Msg("Failed to execute exploit")
 	}
 
-	exploit := config.Exploit{
+	exploitS := config.Exploit{
 		Name: config.ArgsAttackInstance.ExploitPath,
 		PID:  result.Cmd.Process.Pid,
 	}
-	config.ArgsConfigInstance.Exploits = append(config.ArgsConfigInstance.Exploits, exploit)
+	config.ArgsConfigInstance.Exploits = append(config.ArgsConfigInstance.Exploits, exploitS)
 	if err := config.WriteConfig(); err != nil {
 		logger.Log.Fatal().Err(err).Msg("Failed to write configuration with new exploit")
 	}
@@ -234,55 +161,6 @@ func attack(cmd *cobra.Command, args []string) {
 	if err := result.Cmd.Wait(); err != nil {
 		logger.Log.Error().Err(err).Msg("Exploit process exited with error")
 	}
-}
-
-// Detach detaches the current process from the terminal by re-executing itself
-func Detach() {
-	cmd := exec.Command(os.Args[0], os.Args[1:]...)
-
-	filteredArgs := []string{}
-	for _, arg := range os.Args[1:] {
-		if arg != "--detach" && arg != "-d" {
-			filteredArgs = append(filteredArgs, arg)
-		}
-	}
-	cmd = exec.Command(os.Args[0], filteredArgs...)
-
-	cmd.Stdout = nil
-	cmd.Stderr = nil
-	cmd.Stdin = nil
-	cmd.SysProcAttr = &syscall.SysProcAttr{Setsid: true}
-
-	err := cmd.Start()
-	if err != nil {
-		fmt.Println(logger.Red+"[ERROR]"+logger.Reset+"| Error during detach:", err)
-		os.Exit(1)
-	}
-
-	fmt.Println(logger.Yellow+"[WARN]"+logger.Reset+"| Process detached with PID:", cmd.Process.Pid)
-	os.Exit(0)
-}
-
-// ValidateArgs validates the arguments passed to the program
-func ValidateArgs(args config.ArgsAttack) error {
-	if args.TickTime < 1 {
-		return errors.New("tick time must be at least 1")
-	}
-
-	exploitPath, err := filepath.Abs(args.ExploitPath)
-	if err != nil {
-		return fmt.Errorf("error resolving exploit path: %v", err)
-	}
-
-	if info, err := os.Stat(exploitPath); err == nil && info.Mode()&0o111 == 0 {
-		return errors.New("exploit file is not executable")
-	}
-
-	if _, err := os.Stat(exploitPath); os.IsNotExist(err) {
-		return errors.New("exploit not found in the exploits directory")
-	}
-
-	return nil
 }
 
 // ===== CREATE EXPLOIT FUNCTIONS =====
@@ -331,27 +209,27 @@ func Create(cmd *cobra.Command, args []string) {
 
 // ===== LIST EXPLOIT FUNCTIONS =====
 
-// ListExploit lists all currently running exploits
-func ListExploit(cmd *cobra.Command, args []string) {
+// List lists all currently running exploits
+func List(cmd *cobra.Command, args []string) {
 	err := config.LoadLocalConfig()
 	if err != nil {
 		logger.Log.Error().Err(err).Msg("Error loading local configuration, try running `cookieclient config reset`")
 		return
 	}
 	for i := 0; i < len(config.ArgsConfigInstance.Exploits); {
-		exploit := config.ArgsConfigInstance.Exploits[i]
-		proc, err := os.FindProcess(exploit.PID)
+		exploitS := config.ArgsConfigInstance.Exploits[i]
+		proc, err := os.FindProcess(exploitS.PID)
 		if err != nil || proc == nil {
-			logger.Log.Warn().Str("exploit", exploit.Name).Msg("Exploit removed due to invalid process")
+			logger.Log.Warn().Str("exploit", exploitS.Name).Msg("Exploit removed due to invalid process")
 			config.ArgsConfigInstance.Exploits = append(config.ArgsConfigInstance.Exploits[:i], config.ArgsConfigInstance.Exploits[i+1:]...)
 			continue
 		}
 		err = proc.Signal(syscall.Signal(0))
 		if err != nil {
-			logger.Log.Warn().Str("exploit", exploit.Name).Msg("Exploit removed due to inactive process")
+			logger.Log.Warn().Str("exploit", exploitS.Name).Msg("Exploit removed due to inactive process")
 			config.ArgsConfigInstance.Exploits = append(config.ArgsConfigInstance.Exploits[:i], config.ArgsConfigInstance.Exploits[i+1:]...)
 		} else {
-			logger.Log.Info().Str("exploit", exploit.Name).Int("pid", exploit.PID).Msg("Available exploit")
+			logger.Log.Info().Str("exploit", exploitS.Name).Int("pid", exploitS.PID).Msg("Available exploit")
 			i++
 		}
 	}
@@ -402,17 +280,17 @@ func Remove(cmd *cobra.Command, args []string) {
 
 // ===== STOP EXPLOIT FUNCTIONS =====
 
-// stopExploit stops a running exploit by PID
-func stopExploit(cmd *cobra.Command, args []string) {
+// Stop stops a running exploit by PID
+func Stop(cmd *cobra.Command, args []string) {
 	if err := config.LoadLocalConfig(); err != nil {
 		logger.Log.Error().Err(err).Msg("Error loading local configuration, try running `cookieclient config reset`")
 		return
 	}
 
-	for i, exploit := range config.ArgsConfigInstance.Exploits {
-		logger.Log.Info().Str("Exploit", exploit.Name).Msg("Stopping exploit")
-		if exploit.PID == pid {
-			proc, err := os.FindProcess(exploit.PID)
+	for i, exploitS := range config.ArgsConfigInstance.Exploits {
+		logger.Log.Info().Str("Exploit", exploitS.Name).Msg("Stopping exploit")
+		if exploitS.PID == pid {
+			proc, err := os.FindProcess(exploitS.PID)
 			if err != nil {
 				logger.Log.Error().Err(err).Msg("Error finding process")
 				return
@@ -424,7 +302,7 @@ func stopExploit(cmd *cobra.Command, args []string) {
 			}
 			config.ArgsConfigInstance.Exploits = append(config.ArgsConfigInstance.Exploits[:i], config.ArgsConfigInstance.Exploits[i+1:]...)
 			config.WriteConfig()
-			logger.Log.Info().Str("Exploit", exploit.Name).Msg("Exploit stopped successfully")
+			logger.Log.Info().Str("Exploit", exploitS.Name).Msg("Exploit stopped successfully")
 			return
 		}
 	}
