+++ client/internal/exploit/cleanup.go
@@ -1,0 +1,21 @@
+package exploit
+
+import (
+	"slices"
+
+	"github.com/ByteTheCookies/cookieclient/internal/config"
+	"github.com/ByteTheCookies/cookieclient/internal/logger"
+)
+
+// Cleanup removes an exploit from the configuration
+func Cleanup(pid int) {
+	for i, exploit := range config.ArgsConfigInstance.Exploits {
+		if exploit.PID == pid {
+			config.ArgsConfigInstance.Exploits = slices.Delete(config.ArgsConfigInstance.Exploits, i, i+1)
+			break
+		}
+	}
+	if err := config.WriteConfig(); err != nil {
+		logger.Log.Error().Err(err).Msg("Failed to write configuration after exploit cleanup")
+	}
+}
