+++ client/internal/exploit/detach.go
@@ -1,0 +1,37 @@
+package exploit
+
+import (
+	"fmt"
+	"os"
+	"os/exec"
+	"syscall"
+
+	"github.com/ByteTheCookies/cookieclient/internal/logger"
+)
+
+// Detach detaches the current process from the terminal by re-executing itself
+func Detach() {
+	cmd := exec.Command(os.Args[0], os.Args[1:]...)
+
+	filteredArgs := []string{}
+	for _, arg := range os.Args[1:] {
+		if arg != "--detach" && arg != "-d" {
+			filteredArgs = append(filteredArgs, arg)
+		}
+	}
+	cmd = exec.Command(os.Args[0], filteredArgs...)
+
+	cmd.Stdout = nil
+	cmd.Stderr = nil
+	cmd.Stdin = nil
+	cmd.SysProcAttr = &syscall.SysProcAttr{Setsid: true}
+
+	err := cmd.Start()
+	if err != nil {
+		fmt.Println(logger.Red+"[ERROR]"+logger.Reset+"| Error during detach:", err)
+		os.Exit(1)
+	}
+
+	fmt.Println(logger.Yellow+"[WARN]"+logger.Reset+"| Process detached with PID:", cmd.Process.Pid)
+	os.Exit(0)
+}
